+++ packages/wallet/src/initialization/instances/gas-fee-controller/gas-fee-controller.test.ts
@@ -211,6 +211,10 @@ function getInstanceOptions(): WalletOptions['instanceOptions'] {
       clientId: 'test',
     },
     networkController: {
+      analyticsOptions: {
+        isRpcEndpointUrlPublic: (): boolean => false,
+        rpcServiceEventsSampleRate: 0,
+      },
       infuraProjectId: 'test-infura-project-id',
     },
     storageService: {